bot/pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = "Autozakaz bot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.20.0.post0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]
